+++ src/components/App/AppTabs.vue
