+++ src/components/Sura/SuraText.vue
