+++ src/components/Tarteel/TarteelDetail.vue
