+++ src/components/Tarteel/TarteelSearchResults.vue
