--- src/components/TarteelDrawerList.vue
@@ -1,83 +1,0 @@
-<template>
-  <v-list-item
-    v-for="(item, index) in items"
-    :key="index"
-    :value="index"
-    color="grey"
-    :class="{
-      'active-tarteel-item': selectedItem === index,
-    }"
-    @click="$emit('selecteItem', index)"
-  >
-    <template v-slot:prepend>
-      <v-list-item-title class="ps-6">
-        <v-icon :icon="itemIcon" size="20" class="me-2 count-key-item" />
-        {{ item.word }}
-      </v-list-item-title>
-    </template>
-
-    <template v-slot:append>
-      <!-- <span class="text-caption ms-2"> {{ item.count }} </span> -->
-      <v-badge
-        :content="item.count"
-        :color="item.coutingType"
-        class="word-count"
-        offset-x="5"
-        offset-y="0"
-        inline
-      ></v-badge>
-      <!-- <v-badge
-        :content="`${getCoutingTypeLabel(item.coutingType)}`"
-        :color="item.coutingType"
-        class="word-count"
-        offset-x="5"
-        offset-y="0"
-        inline
-      ></v-badge> -->
-      <v-btn
-        icon="mdi-close"
-        variant="text"
-        size="x-small"
-        @click.stop="$emit('deleteItem', index)"
-      ></v-btn>
-    </template>
-  </v-list-item>
-</template>
-
-<script setup>
-defineProps({
-  items: {
-    type: Array,
-    required: false,
-  },
-  selectedItem: {
-    type: Number,
-    required: false,
-  },
-  badgeContent: {
-    type: String,
-    required: false,
-  },
-  itemIcon: {
-    type: String,
-    required: false,
-  },
-})
-
-defineEmits(["selecteItem", "deleteItem"])
-
-const getCoutingTypeLabel = (coutingType) => {
-  if (coutingType === "word-count") {
-    return "مشتق"
-  }
-  if (coutingType === "verse-count") {
-    return "آية"
-  }
-  return ""
-}
-</script>
-<style scoped>
-.v-list-item-title {
-  font-size: 14px;
-}
-</style>
