+++ src/components/Verse/VerseCardItem.vue
