+++ src/components/Verse/VerseDetails.vue
