+++ src/components/Verse/VerseDetailsCard.vue
