+++ src/components/Verse/VerseDetailsWords.vue
