+++ src/components/Verse/VerseNumbers.vue
