+++ src/components/Verse/VerseStats.vue
